pyproject.toml
[tool.poetry]
name = "codeagi"
version = "0.1.0"
description = ""
authors = ["Brian Jinwright <bjinwright@qwigo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
envs = "^1.4"
googlesearch-python = "^1.2.3"
beautifulsoup4 = "^4.12.2"
requests = "^2.28.2"
pinecone-client = "^2.2.1"
langchain = "^0.0.139"
python-dotenv = "^1.0.0"
selenium = "^4.8.3"
unstructured = "^0.5.12"
tiktoken = "^0.3.3"
tqdm = "^4.65.0"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
